crash: move a few code bits to setup support of crash hotplug
The crash hotplug support leans on the work for the kexec_file_load()
syscall. To support crash hotplug for the kexec_load() syscall, a few
bits of code need to be move outside of CONFIG_KEXEC_FILE.

The crash_prepare_elf64_headers() and crash_exclude_mem_range()
functions, as well as the struct crash_mem are needed for support
of crash hotplug on the kexec_load() path, as well as kexec_file_load().
As such, need to move these bits out of kexec_file.c and into a
common location crash_core.c.

No functionality change.

Signed-off-by: Eric DeVolder <eric.devolder@oracle.com>